lib: Add .repeat(), .map(), .sort() and custom action signatures
These two common functions will be useful as built-in
actions.  Repeat takes and action name and executes it
multiple times with the given options.  Results are
returned as a collection.  Map will take a collection
(or lookup reference) as input and execute the given
function on all items, and returning a new array of
the values returned by each iteration.

This change also allows for custom action signatures if
a different style is preferred over `(opts, label)`.
When choosing a custom signature, it will require the
developer to expose their own action function rather
than rely on the auto-expose functionality of
the constructor.

Semver: minor