feat(chain): allow moment object to pass through lookup functions
Moment object normally identifies as 'object', in which case the lookup
function creates a new object and replaces keys with new values from
additional lookup. In the case of moment, this breaks the prototype
chain and further function calls will not work.

This allows for things that look like a moment object to identify as
`moment` and bypass the iteration step

Semver: minor