fix(parser): rework syntax and relax parser for arbitrary values
Previously the parser was being applied to all values sent
through the chain `lookup` method. The syntax was originallyty designed
to support this by including tokens and parsing support for arbitrary
scaler values. In practice, this caused problems with more complex
string patterns in cases where the tokenization was ambiguous.

Instead, the parser and the chain `lookup` method have been updated to
support only `!functions` and `#lookup.values` as explicit parts of the
syntax. All other values will be treated as literals and pass through
unmodified.

Semver: patch
Ref: LOG-8436