feat(actions): include a serial action
Similar to the repeat action be ensure each task is both run and
completes before moving on to the next. With a lot of async tasks the
implementation of repeat can't ensure the order of each task. While it
may ensure each task is invoked in order, it does not ensure each
completes.