feat(chain): add support for luxon object during lookups
With the deprecation of moment, luxon is sane replacement in testing.
In a similar fashion as with moment, we don't want to break the
prototype chain when passing luxon objects around. Unlike moment, there
are several date-ish object we need to acount for. This will allow users
to reference and manimulat luxon object during chain execution with out
having to cast between native date objects